fix: remove `typesVersions` field
adding config and type test for different version of TypeScript.
