+++ .changeset/fresh-spiders-invite.md
@@ -1,0 +1,6 @@
+---
+"type-plus": patch
+---
+
+Removing incorrect usage of the `typesVersions` field in `package.json`.
+
