+++ .vscode/ltex.dictionary.en-US.txt
@@ -8,3 +8,4 @@ NONINFRINGEMENT
 endofunctor
 endofunction
 Endofunction
+bigint
