+++ .vscode/ltex.hiddenFalsePositives.en-US.txt
@@ -5,3 +5,4 @@
 {"rule":"MORFOLOGIK_RULE_EN_US","sentence":"^\\Q[Semantic Release]semantic-release-url\\E$"}
 {"rule":"MORFOLOGIK_RULE_EN_US","sentence":"^\\Q[Visual Studio Code]vscode-url [Wallaby.js]wallaby-url\\E$"}
 {"rule":"EN_US_SIMPLE_REPLACE","sentence":"^\\Q⚗️ :alembic:\n🔄 :arrows_counterclockwise:\n🚛 :articulated_lorry:\n➰ :curly_loop:\n💱 :currency_exchange:\n🌀 :cyclone:\n💠 :diamond_shape_with_a_dot_inside:\n➿ :loop:\n🪄 :magic_wand:\n🔁 :repeat:\n🔀 :twisted_rightwards_arrows:\n:dizzy:\\E$"}
+{"rule":"EN_US_SIMPLE_REPLACE","sentence":"^\\Q⚗️ :alembic:\n🔄 :arrows_counterclockwise:\n🚛 :articulated_lorry:\n➰ :curly_loop:\n💱 :currency_exchange:\n🌀 :cyclone:\n💠 :diamond_shape_with_a_dot_inside:\n➿ :loop:\n🪄 :magic_wand:\n🔁 :repeat:\n🔀 :twisted_rightwards_arrows:\\E$"}
