+++ packages/type-plus/src/array/array.entries.spec.ts
@@ -1,18 +1,5 @@
-import { it, test } from '@jest/globals'
-
+import { it } from '@jest/globals'
 import { type ArrayPlus, testType } from '../index.js'
-
-test('behavior of array.entries()', () => {
-	const array = [1, 2, '3']
-	const entries = array.entries()
-	testType.equal<typeof entries, IterableIterator<[number, string | number]>>(true)
-})
-
-test('behavior of tuple.entries()', () => {
-	const tuple = [1, 2, '3'] as const
-	const entries = tuple.entries()
-	testType.equal<typeof entries, IterableIterator<[number, 1 | 2 | '3']>>(true)
-})
 
 it('gets Array<[number, T]> for array', () => {
 	testType.equal<ArrayPlus.Entries<string[]>, Array<[number, string]>>(true)
