+++ packages/type-plus/src/array/array_plus.find.ts54.spec.ts
@@ -1,0 +1,8 @@
+import { test } from '@jest/globals'
+import { testType } from '../index.js'
+
+test('behavior of array.find()', () => {
+	const array = [1, 2, '3']
+	const r = array.find((x) => typeof x === 'number')
+	testType.equal<typeof r, number | string | undefined>(true)
+})
