+++ packages/type-plus/tsconfig.json
@@ -1,5 +1,5 @@
 {
-	"extends": "./tsconfig.esm.json",
+	"extends": "./tsconfig.ts56.json",
 	"typedocOptions": {
 		"entryPoints": ["src/index.ts"]
 	}
