+++ packages/type-plus/tsconfig.ts54.json
@@ -1,0 +1,4 @@
+{
+	"extends": "./tsconfig.esm.json",
+	"exclude": ["node_modules", "src/type-checker", "src/types", "src/**/*.ts55.spec.ts", "src/**/*.ts56.spec.ts"]
+}
