+++ pages/astro.config.mjs
@@ -18,7 +18,7 @@ export default defineConfig({
 					label: 'Guides',
 					items: [
 						// Each item here is one entry in the navigation menu.
-						{ label: 'Example Guide', link: '/guides/example/' },
+						{ label: 'Getting Started', link: '/guides/getting-started/' },
 					],
 				},
 				{
