+++ pages/src/content/docs/guides/getting-started.mdx
@@ -1,0 +1,31 @@
+---
+title: Getting Started
+---
+
+[type-plus] provides a set of utility types and functions for applications, library, and type-level programming.
+
+## Installation
+
+```sh
+npm install type-plus
+
+yarn add type-plus
+
+pnpm add type-plus
+```
+
+## TypeScript Version Compatibility
+
+[TypeScript] version does not follow the [Semantic Versioning] specification.
+That means every release of TypeScript may have breaking changes.
+
+Being a type library, especially a type library with a lot of advanced types,
+[type-plus] is at the forefront of this battleground.
+
+Starting from version 8.0.0,
+[type-plus] will be more compatible with various versions of TypeScript.
+Depends on which typescript you are using, [type-plus] will be loaded with a set of types that are compatible with that version of TypeScript.
+
+[type-plus]: https://github.com/unional/type-plus
+[TypeScript]: https://www.typescriptlang.org/
+[Semantic Versioning]: https://semver.org/
